third_party/WebKit/LayoutTests/fast/dom/navigator-vibration-expected.txt

Test for the Vibration API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL navigator.vibrate(0); should be undefined. Threw exception TypeError: Object #<Navigator> has no method 'vibrate'
FAIL navigator.vibrate([]); should be undefined. Threw exception TypeError: Object #<Navigator> has no method 'vibrate'
FAIL navigator.vibrate(1000); should be undefined. Threw exception TypeError: Object #<Navigator> has no method 'vibrate'
FAIL navigator.vibrate([1000, 300, 500]); should be undefined. Threw exception TypeError: Object #<Navigator> has no method 'vibrate'
FAIL navigator.vibrate(4294967295); should be undefined. Threw exception TypeError: Object #<Navigator> has no method 'vibrate'
PASS navigator.vibrate(); threw exception TypeError: Object #<Navigator> has no method 'vibrate'.
PASS successfullyParsed is true

TEST COMPLETE
